covid_data: fix regression in 14 day deltas
Because of a regression introduced in 809d543dd226e5623b9d84ae8bcb01e323beb04b,
the delta in the 14 day rolling sum was just ignoring
the days that dropped off since the last update, so
it would only go up.

To fix this, we need to compute old_rolling from the last day
we had data.